% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pprop.R
\name{pp.pprop}
\alias{pp.pprop}
\alias{print.pprop}
\title{Pretty print a \code{pprop} object.}
\usage{
pp.pprop(x, signif.digits = getOption("pp.signif.digits"),
  round.digits = getOption("pp.round.digits"),
  format.params = getOption("pp.format.params"), ...)

\method{print}{pprop}(x, signif.digits = getOption("pp.signif.digits"),
  round.digits = getOption("pp.round.digits"),
  format.params = getOption("pp.format.params"), ...)
}
\arguments{
\item{x}{Output of \code{link{pprop}} when \code{char = FALSE}.}

\item{signif.digits}{See \code{link{pp}}.}

\item{round.digits}{See \code{link{pp}}.}

\item{format.params}{See \code{link{pp}}.}

\item{...}{Ignored.}
}
\value{
\code{pp} prints \code{x}. Do not rely on the value, if any, that \code{pp} might return.
}
\description{
Pretty print a \code{pprop} object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Alias for \code{pp.pprop}.
}}
\seealso{
\code{\link{pp}}, \code{\link[base]{print}}.

Other pretty.print.functions: \code{\link{pp}}
}
\keyword{print}
\keyword{utilities}

