% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp}
\alias{pp}
\title{Pretty print any object.}
\usage{
pp(x = NA, both = FALSE, digits = getOption("pp.digits"),
  signif.digits = getOption("pp.signif.digits"),
  round.digits = getOption("pp.round.digits"),
  format.params = getOption("pp.format.params"))
}
\arguments{
\item{x}{An object to print.}

\item{both}{Show both, the \code{print} (unformatted) version and the formatted version of \code{x}? \code{FALSE} (default) only shows the formatted version.}

\item{digits}{For \code{x} between -1 and 1, this is the same as \code{round.digits}. For \code{x} outside of the -1 to 1 range, same as \code{signif.digits}. \code{NA} (default) = ignore.}

\item{signif.digits}{Integer indicating the number of significant digits to be used, or \code{NA}. Ignored if \code{digits} or \code{round.digits} is not \code{NA}. See \code{\link[base]{signif}}.}

\item{round.digits}{Integer indicating the number of decimal places to be used, or \code{NA}. \code{NA} (default) = ignore. Ignored if \code{digits} is not \code{NA}. See \code{\link[base]{round}}.}

\item{format.params}{Parameters passed to \code{\link[base]{format}}.}
}
\value{
\code{pp} prints \code{x}. Do not rely on the value, if any, that \code{pp} might return.
}
\description{
Print values in a way that humans can easily understand.
}
\details{
At its core, \code{pp} is a wrapper around \code{\link[base]{signif}}, \code{\link[base]{round}}, and \code{\link[base]{format}}. Currently, methods are defined for the following classes: 
\itemize{
	\item integer, numeric
\item matrix, array
	\item data.frame (optionally, the ``digits'' parameters can be vectors, in which case a different value applies to each column of the data.frame)
	\item table (output from \code{\link[base]{table}} and \code{\link[stats]{xtabs}})
	\item lm, glm (prints odds ratios for logistic regression)
	\item AsIs (output from \code{\link[base]{I}})
	\item list
	\item pprop (output from \code{\link{pprop}} -- ``pretty proportion'')
}

If a ``digits'' parameter is \code{NA}, it is ignored. At most, only one non-NA ``digits'' parameter is used. Their order of precedence is: \code{digits}, \code{round.digits}, \code{signif.digits}. By default, \code{digits} and \code{round.digits} are \code{NA}.
}
\examples{
## Create a dataset
set.seed(42)
x.big = rnorm(12, mean=1e4, sd=1e3)
x.small = rnorm(12, mean=0, sd=0.5)
name = sample(levels(iris$Species), 12, replace=TRUE)
dd = data.frame(name, x.big, x.small)

pp(dd)			# typical usage
pp(dd, b=TRUE)	# show original and formatted

pp(dd, s=2)		# 2 significant digits
pp(dd, r=2)		# 2 decimal places
# 2 decimal places for small numbers (-1 to 1), 2 significant digits otherwise
pp(dd, d=2)	

pp(dd, s=NA)	# format only, don't round
pp(dd, r=-3)	# round to the nearest thousand

# different signif.digits and round.digits for each column
pp(dd, s=c(NA, NA, 3), r=c(NA, 2, NA))

mm = tcrossprod(x.big, x.small)
pp(mm)	# matrix
pp(qr(mm))	# list

flowers = sample(levels(iris$Species), 1e4, replace=TRUE)
pp(table(flowers))	# table

## Linear Models -- example from lm
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
lm.D9 <- lm(weight ~ group)

summary(lm.D9)
pp(lm.D9)	# cleaner output

## Logistic regression
## Create a dataset
set.seed(42)
n = 100
dd = data.frame(x1 = runif(n), x2 = runif(n))
dd = within(dd, {
		x3 = 2 * x1 + x2 + rnorm(n)
		A = factor(cut(x1, 3, FALSE))
		B = factor(cut(x2, 3, FALSE))
		y = factor(cut(x3, 2, FALSE))
	})

gg = glm(y ~ A + B, data = dd, family = "binomial")
summary(gg)	
pp(gg)	# Odds ratios, 95\% CI's
}
\seealso{
\code{\link[base]{print}}, \code{\link[base]{signif}}, \code{\link[base]{round}}, \code{\link[base]{format}}, \code{\link[base]{summary}}.

Other pretty.print.functions: \code{\link{pp.pprop}}
}
\keyword{print}
\keyword{utilities}

