% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyprint.R
\docType{package}
\name{prettyprint-package}
\alias{prettyprint-package}
\title{prettyprint: A suite of functions that print values in a way that humans can 
easily understand.}
\description{
prettyprint: A suite of functions that print values in a way that humans can 
easily understand.
}

