% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp.R
\name{pp}
\alias{pp}
\title{Pretty print any object.}
\usage{
pp(x = NA, both = FALSE, signif.digits = getOption("pp.signif.digits"),
  round.digits = getOption("pp.round.digits"),
  format.params = getOption("pp.format.params"))
}
\arguments{
\item{x}{An object to print.}

\item{both}{Show both, the \code{print} (unformatted) version and the formatted version of \code{x}? FALSE (default) only shows the formatted version.}

\item{signif.digits}{Integer indicating the number of significant digits to be used, or NA. Ignored if round.digits is not NA. See \code{\link[base]{signif}}.}

\item{round.digits}{Integer indicating the number of decimal places to be used, or NA. NA (default) means that \code{signif.digits} is used. See \code{\link[base]{round}}.}

\item{format.params}{Parameters passed to \code{\link[base]{format}}.}
}
\value{
\code{pp} prints \code{x}. Do not rely on the value, if any, that \code{pp} might return.
}
\description{
Print values in a way that humans can easily understand.
}
\details{
At its core, \code{pp} is a wrapper around \code{\link[base]{signif}}, \code{\link[base]{round}}, and \code{\link[base]{format}}. Currently, methods are defined for the following classes: 
\itemize{
	\item integer, numeric
\item matrix, array
	\item data.frame (optionally, \code{signif.digits} and \code{round.digits} parameters can be vectors, in which case a different value applies to each column of the data.frame)
	\item table (output from \code{\link[base]{table}} and \code{\link[stats]{xtabs}})
	\item lm, glm (prints odds ratios for logistic regression)
	\item AsIs (output from \code{\link[base]{I}})
	\item list
	\item pprop (output from \code{\link{pprop}} -- ``pretty proportion'')
}
}
\examples{
## Create a dataset
set.seed(42)
x.big = rnorm(12, mean=1e4, sd=1e3)
x.small = rnorm(12, mean=1e-3, sd=1e-3)
name = sample(levels(iris$Species), 12, replace=TRUE)
dd = data.frame(name, x.big, x.small)

pp(dd)	# typical usage
pp(dd, b=TRUE)	# show original and formatted

flowers = sample(levels(iris$Species), 1e4, replace=TRUE)
pp(table(flowers))	# table

pp(dd, s=NA)	# format only, don't round
pp(dd$x.big, r=2)	# show 2 decimal places
# different signif.digits and round.digits for each column
pp(dd, s=c(NA, NA, 3), r=c(NA, 2, NA))

mm = tcrossprod(x.big, x.small)
pp(mm)	# matrix
pp(qr(mm))	# list

## Linear Models -- example from lm
ctl <- c(4.17,5.58,5.18,6.11,4.50,4.61,5.17,4.53,5.33,5.14)
trt <- c(4.81,4.17,4.41,3.59,5.87,3.83,6.03,4.89,4.32,4.69)
group <- gl(2, 10, 20, labels = c("Ctl","Trt"))
weight <- c(ctl, trt)
lm.D9 <- lm(weight ~ group)

summary(lm.D9)
pp(lm.D9)	# cleaner output

## Logit Regression -- example from http://statistics.ats.ucla.edu/stat/r/dae/logit.htm
mydata <- read.csv("http://www.ats.ucla.edu/stat/data/binary.csv")
mydata$rank <- factor(mydata$rank)
mylogit <- glm(admit ~ gre + gpa + rank, data = mydata, family = "binomial")

summary(mylogit)
pp(mylogit)	# Odds ratios, 95\% CI's
}
\seealso{
\code{\link[base]{print}}, \code{\link[base]{signif}}, \code{\link[base]{round}}, \code{\link[base]{format}}, \code{\link[base]{summary}}.

Other pretty.print.functions: \code{\link{pp.pprop}}
}
\keyword{print}
\keyword{utilities}

